Added rendering mesh lines. Cosmetic changes.
